Move ND ops into nddualnumber and clean up includes
